Art assets for laboratory stage boss
The images were downloaded from craftpix.net. It is a combination of multiple freebies offered by the website and not all of them are actually used in the game.

I did prefer to keep them in because they offer nice ways of expanding the functionality and aesthetic of the game if I or someone else decides to refurbish it once again.

https://craftpix.net/file-licenses/