Update Player.java
The most important thing which I was trying to do is fix the wall clipping. I did not completely succeed but it is certainly better than it was.
That means that a lot of changes were made in setPositionOfPlayer() and onGround(...).

Besides that... COMMENTS, COMMENTS, COMMENTS as I was trying to figure out what it does since it was mostly written by Nathan