Update GUI.java
Loads of changes.

Most importantly I made it possible to win the game. That required:
Implementation of collision with the boss and some extra animations.

The boss was not fully developed, so I hope I made it more manageable and clearer. Also now the boss shoots, its bullets collide. The boss can die which leads to winning the game.

Added appropriate backgrounds and tired to add more meaningful comments.